Updated code and tests to check for non-existing select items from list
Now check to see if select list is multi-selection or not.  If it is multi-
select then raise error for any non-existing item(s). If it is a single
select list then log warning if any item, up to the second to last item, that
does not exist within the select list. And if the last item does not exist
then a ValueError is thrown.  This discerning error/warning reporting for
single select list was designed to accommodate current stated operation of the
keyword, "If more than one value is given for a single-selection list, the
last value will be selected."