Register browser even if get url fails.
An issue was reported with SeleniumLibrary [1] where if "Open Browser"
succeeds in starting the browser but the get url call fails the opened
browser doesn't get registered with the Library's internal browser registery.
Thus when 'Close All Browsers' is called it fails to close the opened browser.
This patch fixes that problem.

[1] http://code.google.com/p/robotframework-seleniumlibrary/issues/detail?id=220