Allow keywords to use WebDriver Element in addition to Locator.
In some cases I want to be able to store the element itself and then act upon
the element as compared to passing a locator each time and re-finding the
element. At this point I am implementing this by checking the locator argument
within keywords/_elements.py. This could possibly fit into the element finder
code but this approach seems more desirable in that it is the point where the
"locator" is passed in and an element is returned. Now instead it checks to
see if the "locator" is a string and if so tries to find the element. And if
it is an element is goes ahead and passes it through. Testing is needed to verify this doesn't break anything in addition to verify no other keywords would have issues if passed a WebDriver Element instead of a locator
string.