Removed 'Get WebElement' (singular) keyword.
Removing the placeholder for the (proposed) singular 'Get WebElement' keyword.
In discussions with Kevin and Pekka it was suggested that the having only
'Get WebElements' (multiple) would provide both cases and help keep number
of keywords to a reasonible number. In addition, it has been suggested to
add either a index or firt-only flag to 'Get WebElements' (to come in
another commit).