(Properly) display Firefox version in Travis.ci output
Trying to display version information in travis ci build. Removed previous attempt and brought back old code used to do this. Trying to see which Firefox version we are running to know why tests are failing. There is known issues with running the very latest Firefox version when selenium has not yet be updated.