Added possibility to combine index and filename args
If test are run, to get a custom and unique filename
each time screenshot is takes, it would require
creating blobal variable in RF data and making
sure that variable in unique for each time
screenshot is taken. This is now much easier
to do with the index param in the Capture
Page Screenshot keyword.