Selenium2Library new architecture (#770)
This PR removes the library multi inheritance madness. This change result cleaner code base, which eases the maintenance and hopefully eases accepting the new enhancement to the library.

The PR also removes the meta classes and the automatic keyword decoration in the keywordgroup object. Also the PR introduces a new context object(which is the library itself). The context object holds information from the different WebDriver instances and allows methods to register and fetch the different WebDriver instances