Support `:` as locator separator in addtion to `=`. See #908.
Still need to decided how to handle locators which just happen to have
`:` in them like `foo:bar`. With the current implementation they fail
with a ValueError becaue `foo` is not a valid locator prefix. Could
also change the logic to handle this so that `:` and `=` are
separators *only* if the part before is a valid prefix.

Docs are missing too.