Alternative locator syntax (#909)
* Support `:` as locator separator in addtion to `=`. See #908.

Still need to decided how to handle locators which just happen to have
`:` in them like `foo:bar`. With the current implementation they fail
with a ValueError becaue `foo` is not a valid locator prefix. Could
also change the logic to handle this so that `:` and `=` are
separators *only* if the part before is a valid prefix.

* Remove dead and borken code

* Remove duplicate unit tests.

Didn't notice earlier that _parse_locator was already tested. Decided
to remove old tests because the ones I added are a little simpler due
to the use of `_verify_parse_locator` helper. Added few more tests to
cover everything the old tests covered.

* Use locator prefix only if it is valid.

Earlier `foo=bar` caused ValueError if `foo` wasn't a valid
locator. Now this is same as `default=foo=bar`.

Main motivation to the change is avoiding bwic problems now when we
are adding support for locators using`foo:bar` syntax (#908).