Browser management doc and message update (#952)
* Clean up window tests.

Also add explicit tests for using "Set Window Position/Size" with
strings. Selenium seems to already convert values passed to it to
integers, decided to add conversion also to these keywords so that we
don't depent on that behavior.


- Move default locator strategy docs to beginning of the whole locator
  strategy section. This is the most common approach.
- Move note about old limitation with tables to a new place.

Documentation cleanup is related to #925.

Select Frame and Select Window documentation still todo.