Initial commit of additional patching for selenium 3 move_to bug.
First, I'll note my workaround is slightly different. Instead of catching attribute errors,
I pass in both the driver and the element(s) checking first if this is indeed an
EventFiringWebDriver and not Selenium 4 or greater. I did this because I didn't see the other
try/catch method till after I wrote this. I can modify these to go that path. My one concern is
that in all the additional places discovered which are affected I was worried that we might
need multiple try/catch within a single method or one a large block of code. The other concern
was that we may get a AttributError for other reasons.

I added in acceptance tests (but no unit tests yet) for the additional changes. I copied pasted
into the current selenium_move_to_workaround.robot test suite. This feels a little hackish to me
and I would like to either copy paste into multiple test suites under a folder (still hackish).
Or what I would really like to do is to use the existing test suites and accomplish this with
tags and some mechanism to run with a differnt open browser setup. This may be too complicated
for now (mostly in time to figure out that solution) so open for ideas here ..

I have not yet tested with selenium 4 due to a bug I encountered when running the acceptance tests.
Need to look at that now.

Welcome any feedback...