Use python interpreter that executed atest/run.py instead of python
When running under a virtualenv, using `python` as the interpreter
forgets the virtualenv python which executed atest.run.py. Instead using
sys.executable which will execute with same interpreter used to execute
atest/run.py. I also later check to make sure the interpreter is not
None nor an empty string as doc states [1] it may be if "Python is unable
to retrieve the interpreter".

Fixes #1796

[1] https://docs.python.org/3.10/library/sys.html#sys.executable