Added some more Page Should Contain tests verifing custom loglevel
In verifying the documentation on the `Page Should Contain` keyword, as
outlined in issue #1785, I've added a few addition tests which check out
when the custom log level is set to INFO, WARN, NONE and a level below
the current log level. These test verify the general rules laid out in
the keyword documentation.

What I am currently questioning is the default value of the loglevel
argument and whther or not it is actually None and INFO which the
documentation says it is nor TRACE which is what the code seems to have
as a default. This last piece is what is nost baffling as it does indeed
look like the default value is TRACE.