Added another Page Should Contain test to prove default custom loglevel
Added `Page Should Contain Using Default Custom Log Level` as a means to
test out the default log level and that the html code should be printed
out at that level. The original `Page Should Contain` test case (yes, a
test case whcih has the same name as the keyword) as written now does
not print out the html. Apparently this test suite is running at a log
level of DEBUG. And thus with a default custom log level of TRACE no
html should be printed. But if one set the log level to TRACE, as this
added test case does, and leave the custom log level to its default
value, this proves that default is TRACE.