Minor corrections to the release notes
These will be moved over to the v6.1.0 release. No second release
candidate be made as it was not necessary. I did want to work out the
release notes prior to making the v.6.1.0 release and so they had gone
into this v6.1.0rc2 README as a temporary placeholder.

Thanks to Hélio Guilherme for reviewing and suggesting corrections to
the release notes.