Using simplified usage of each broswers Service class
Instead of a complicated library written Service Class, like for
Options, I went back to a extremely simply workaround to accept the
existing Open Browser Keyword arguments as is and just pass those which
were going to the driver creater method into the service creater and
pass the resulting opject along. Nothing fancy needed. These changes
reflect this.

A couple other changes the once seperate Firefox creation argument of the
profile has been moved to the options argument. This seems like a
straight forward fix.

The last is the dropping of desired capabilities. This has been changing
internal with selenium and anyone previously using desired capabilities
should have noticed this and corrected already. At this point selenium
has simply dropped the argument and we have take it in but not passed it
along. This we will need to document and maybe code a nicer warning than
silence.