Initial commit and sketchings of `Print Page As PDF` Keyword
Some initial ideas and sketches for the `Print Page As PDF` keyword. I
wanted to seewhat it may look like and how it may work. So this is just
some preliminary ideas. Seeing a need to handle Print Options, need to
document options and what are valid values, how to deal with the return
encoded base64 string, etc.