Initial atest for remote browsers without options
This is an initial attempt/design for a test suite that verifies
"remote" browsers work without options. This is to test against the
issue outlined in #1855. I have been manually runnning the webdrivers
locally but this second edition is trying to do that in the setup and
then cleanup in the teardown. For some reason this is not working and I
need to investigate further.