Fixed issue with Get Cookie logging assertions
There was some missed chnages to the log. In addition, Chrome now has a
limit on length of expiry with cookies. Its maximum length is 400 days,
as per, https://developer.chrome.com/blog/cookie-max-age-expires/. As
such back off the far future date to one just about a year out. Starting
to think the hardcoded expiry is better in that one can immediately see
it it verse the calculate some date in the future each time the test is
run. Will revisit this in the future.