Temporarily removed inline flags for ignoring case on log statuschecker
Due to an conflict with statuschecker and startingwith Python 3.11,
inline flags, like `(?i)`, must be at the front of the regex. But
statuschecker (v3.0.1) inserts a match begining of line to the start,
thus violating this rule. Discussion have been had over with the status
checker project on possible changes. So in the mean time going to remove
these from the log status check.