Removed webdriverpath environment variable
As a way to deal with constantly changing selenium-manager logging which
affects our status checking I implemented this change. Works fine with
GitHub CI but has some issues on local. In particular under Windows
build & test it is problematic under virtualenvs. It is possible to
resolve but right now its just too problematic .. so pulling out this
change for now.