Removed `Multiple Locators as a List should work` test case
As written this is an invalid test given current functionality. One can
not give a list directly as a locator. (It is possible to check a list
of elements by using a cascaded locator in that for the end locator to
be found all ther other ones must be there as well. But thi is really a
different type of locator as list check).

I am not opposed to expanding the functionality to include this but as
currently stands I see this as an invilad check on functionality that is
not yet provided.