Allowed for time flexibility within cookie log test to handle timezones
The `Test Get Cookie Keyword Logging` test case has a slight issue in
that Get Cookie returns the expir date and converts to local timezone.
So it makes it difficult to run both within CI and across all our
developers timezones without failing. This change is made to correct for
that. I do see that we could shift this some more so that UTC expiry
keeps within the same day from a California to Australia execution.