Initial retry of creating a Service class
Working on an issue with the Options class gave me deeper insight into how it was working. Brought that experience over to
the Service class. Thjis is an initial implementation / rewrite of the Service class. It follows more closely the options class.
There is just the one test for now and I see I need to bring in the service argument into Open Browser but this feels like
a good start.