Completed initial construct of service class
Although I have completed the intial implementation of the service class and
it's usage, I have discovered unlike the selenium options most if not all of the attributes
must be set on the instantiate of the class. So need to rework the parsing (should check
signature instead of attributes) and then construct the class with these parameters.