Reduced the timeout on the expected timeout within expected conditions
The delay for the title change was 600 miliseconds while the timeout was
0.5 seconds. I suspect this gives little time for error. The 600
milisecond timeout was specifically choosen along with the other
timeouts in the expected conditions tests such that any combination of
delays would be unique. This gives us a framework to test various
combinations. So I didn't want to change that. Noting I could simply
swap the 600 millisecond test for the 900 millisecond tes .. which might
be the best bet. But for now trying to reduce the timeout to 0.3
seconds.