Almost complete buit still an error with instatiating the service class
It is almost complete but still something wrong with how I am instatiating the
service class with the parsed arguments. Cleaned up the Service class and added
a test case where I wasn't trying to set the driver path (which is a bad idea to
use a made up one because the test fails in the the chromedriver wouldn't start).