Removed unit test which checked spaces on service argument
My method for parsing out the argument does not allow for spaces. I want to revist this but
in the meantime going to exclude that test.

Also corrected unit test which checked the plug in documentation. As I update the library
intro section this was failing. Not sure why it was not failing in GitHub Actions .. either
I hadn't pushed or maybe had not yet checked that yet ..