Added unit tests to check for BASE64 and proper handling
There is a still an error when the library is initialized with
something like `bASe64`. It is not setting it. From the code this looks
like the unintended behavior. But then I don't understand why theother test case
works when `EmBed` is given. This too should fail as per my reading of the code.
Going to trace through a bit more to see why that isn't failing and what I may be
missing.