Removed sleep and extended expected timeout.
Previously the timeout was 1 sec and the alert was delayed 500ms. With a normal execution
it seems that that 500ms diference is just not long enough to reasonably expect the alert
to appear. In reviewing this test it was shown any delay was useful and the needed time
for an alert to triggered and then recognized would need more than 500ms.